requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch == 2.3.1+cpu
fastapi
uvicorn
pydantic
numpy == 1.26.4
pandas
scikit-learn
sentence-transformers == 2.7.0
tqdm
beautifulsoup4
requests
pdfminer.six
streamlit
openai
python-dotenv